HP-100 : Product response 폴더 구조 리팩토링
