+++ src/main/java/com/happypill/application/service/product/ProductService.java
@@ -7,9 +7,9 @@ import com.happypill.application.exception.custom.ExceptionCode;
 import com.happypill.application.exception.global.BusinessException;
 import com.happypill.application.repository.product.ProductRepository;
 import com.happypill.application.repository.productprice.ProductPriceRepository;
-import com.happypill.application.service.product.dto.response.CustomPageResponse;
-import com.happypill.application.service.product.dto.response.ProductInfoResponse;
-import com.happypill.application.service.product.dto.response.ProductResponse;
+import com.happypill.application.service.product.response.CustomPageResponse;
+import com.happypill.application.service.product.response.ProductInfoResponse;
+import com.happypill.application.service.product.response.ProductResponse;
 import com.happypill.application.service.product.response.ProductRelatedResponse;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
