+++ src/test/java/com/happypill/application/service/product/ProductServiceTest.java
@@ -12,9 +12,9 @@ import com.happypill.application.repository.categoryinfo.CategoryInfoRepository;
 import com.happypill.application.repository.product.ProductRepository;
 import com.happypill.application.repository.productinfo.ProductInfoRepository;
 import com.happypill.application.repository.productprice.ProductPriceRepository;
-import com.happypill.application.service.product.dto.response.CustomPageResponse;
-import com.happypill.application.service.product.dto.response.ProductInfoResponse;
-import com.happypill.application.service.product.dto.response.ProductResponse;
+import com.happypill.application.service.product.response.CustomPageResponse;
+import com.happypill.application.service.product.response.ProductInfoResponse;
+import com.happypill.application.service.product.response.ProductResponse;
 import com.happypill.application.service.product.response.ProductRelatedResponse;
 import com.happypill.application.util.SnowflakeUtil;
 import jakarta.transaction.Transactional;
